requirements.txt
loguru~=0.5.3
numpy~=1.22.3
torchvision~=0.11.2
tensorboardx~=2.2
librosa~=0.9.1
scikit-learn~=1.1.1
spafe~=0.1.2